% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atwater.R
\name{Atwater}
\alias{Atwater}
\title{Atwater Conversion}
\usage{
Atwater(carb, fat, protein, id, data)
}
\arguments{
\item{carb}{Carbohydrate/TNC values on a DM basis. Atwater factor = 4}

\item{fat}{Fat/Lipid values on a DM basis. Atwater factor = 9}

\item{protein}{Protein/CP values on a DM basis. Atwater factor = 4}

\item{id}{an optional argument. Grouping variable or other indicator of multiple groups within your dataset.}

\item{data}{the dataframe holding your data}
}
\value{
dataframe of ME values from DM values, based on general Atwater factors
}
\description{
This function converts fat, carbohydrate, and protein values from a DM basis to metabolizable energy based on general Atwater factors
All you need to do is input 3 nutrients and a grouping variable, and voila!
This function can be used to convert/prepare your data for further geometric analysis
}
\examples{
Atwater(data=gorillafood,carb=TNC,fat=LIP,protein=CP,id=GROUP)
}
