% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{Online}
\alias{Online}
\title{Online Balancer Wrapper}
\description{
Wraps a balancer to automatically double the sample size N if the original
limit is exceeded.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{balancer}}{The underlying balancer object.}

\item{\code{cls}}{The class generator (e.g., BWD or MultiBWD).}

\item{\code{cls_name}}{The string name of the class (for serialization).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{definition}}{Dictionary of definition parameters.}

\item{\code{state}}{Dictionary of current state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Online-new}{\code{Online$new()}}
\item \href{#method-Online-assign_next}{\code{Online$assign_next()}}
\item \href{#method-Online-assign_all}{\code{Online$assign_all()}}
\item \href{#method-Online-update_state}{\code{Online$update_state()}}
\item \href{#method-Online-reset}{\code{Online$reset()}}
\item \href{#method-Online-process_x}{\code{Online$process_x()}}
\item \href{#method-Online-update_imbalance}{\code{Online$update_imbalance()}}
\item \href{#method-Online-update_path}{\code{Online$update_path()}}
\item \href{#method-Online-clone}{\code{Online$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-new"></a>}}
\if{latex}{\out{\hypertarget{method-Online-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Online wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$new(cls, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cls}}{The class generator (e.g., BWD) OR its character string name.}

\item{\code{...}}{Arguments passed to the balancer's initialize method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-assign_next"></a>}}
\if{latex}{\out{\hypertarget{method-Online-assign_next}{}}}
\subsection{Method \code{assign_next()}}{
Assign treatment to the next point, expanding N if necessary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$assign_next(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Covariate profile vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Treatment assignment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-assign_all"></a>}}
\if{latex}{\out{\hypertarget{method-Online-assign_all}{}}}
\subsection{Method \code{assign_all()}}{
Delegate assign_all to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$assign_all(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix of covariate profiles.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-update_state"></a>}}
\if{latex}{\out{\hypertarget{method-Online-update_state}{}}}
\subsection{Method \code{update_state()}}{
Delegate update_state to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$update_state(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{State arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Online-reset}{}}}
\subsection{Method \code{reset()}}{
Delegate reset to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-process_x"></a>}}
\if{latex}{\out{\hypertarget{method-Online-process_x}{}}}
\subsection{Method \code{process_x()}}{
Delegate process_x to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$process_x(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Covariate vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-update_imbalance"></a>}}
\if{latex}{\out{\hypertarget{method-Online-update_imbalance}{}}}
\subsection{Method \code{update_imbalance()}}{
Delegate update_imbalance to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$update_imbalance(x, a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Processed covariate vector.}

\item{\code{a}}{Assignment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-update_path"></a>}}
\if{latex}{\out{\hypertarget{method-Online-update_path}{}}}
\subsection{Method \code{update_path()}}{
Delegate update_path to inner balancer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$update_path(x, a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Covariate vector.}

\item{\code{a}}{Assignment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Online-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Online-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Online$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
