% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Railfunction.R
\name{Railfunction}
\alias{Railfunction}
\title{Nutrient rails}
\usage{
Railfunction(data, id, x, y)
}
\arguments{
\item{data}{the dataframe holding your data}

\item{id}{your grouping variable or sample ID}

\item{x}{X axis nutrient (will be labelled on an energy (kCal) basis)}

\item{y}{Y axis nutrient (will be labelled on an energy (kCal) basis)}
}
\value{
A complete plot demonstrating the intake rail for ALL ENTRIES (see Groupwiserail for rails by id variable), as well as the ratio of Nutrient X to Y
}
\description{
This function generates a rail plot demonstrating the nutrient balance of foods. 
All you need to do is input 2 nutrients and a grouping variable, and voila!
}
\examples{
Railfunction(baboondailyintake,individual,carb,fat)
}
