% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{serialization}
\alias{serialization}
\alias{serialize_bwd}
\alias{deserialize_bwd}
\title{Serialization Utilities}
\usage{
serialize_bwd(obj)

deserialize_bwd(json_str)
}
\arguments{
\item{obj}{The BWD balancer object to serialize.}

\item{json_str}{A JSON string representing a serialized BWD object.}
}
\description{
Utilities to serialize and deserialize BWD objects to/from JSON.
}
