% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{stop_sample_size_expended}
\alias{stop_sample_size_expended}
\title{Sample Size Expended Error}
\usage{
stop_sample_size_expended()
}
\description{
A custom error condition thrown when the number of units assigned exceeds
the pre-allocated sample size N.
}
