% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xylem_scaling_simulation_dataset.R
\docType{data}
\name{xylem_scaling_simulation_dataset}
\alias{xylem_scaling_simulation_dataset}
\title{Xylem Conduit Scaling Data}
\format{
A data frame with 441,904 rows and 3 variables:
\describe{
\item{Organ}{Plant organ in which xylem conduits were measured (e.g., stem, leaf)}
\item{L}{Distance from leaf tip (m)}
\item{DAVG}{Xylem conduit diameter (µm)}
}
}
\source{
Generated via \code{data-raw/build_xylem_scaling_dataset.R}.
}
\usage{
xylem_scaling_simulation_dataset
}
\description{
A data frame used as an example in \pkg{balancR}, representing simulated xylem conduit measurements.
This data set was used to generate Fig. 3 in \doi{10.1111/geb.70019}.
}
\keyword{datasets}
