% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Railfunction.R
\name{Railfunction}
\alias{Railfunction}
\title{Nutrient rails}
\usage{
Railfunction(data, x, y, id, group = FALSE)
}
\arguments{
\item{data}{the dataframe holding your data}

\item{x}{X axis nutrient (will be labelled on an energy (kCal) basis)}

\item{y}{Y axis nutrient (will be labelled on an energy (kCal) basis)}

\item{id}{an optional argument. Grouping variable or other indicator of multiple groups within your dataset.}

\item{group}{Default is FALSE. Where group=FALSE, intake rail will be generated for entire dataset. Where group=TRUE, intake rail will be generated for each group.}
}
\value{
A complete plot demonstrating the intake rail(s), as well as the ratio of Nutrient X to Y
}
\description{
This function generates a rail plot demonstrating the nutrient balance of foods. 
All you need to do is input 2 nutrients and a grouping variable, and voila!
}
\examples{
Railfunction(baboondailyintake,carb,fat,individualname,group=TRUE)
}
