% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMTfunction.R
\name{RMTfunction}
\alias{RMTfunction}
\title{Right-angle Mixture Triangles}
\usage{
RMTfunction(data, x, y, z, id)
}
\arguments{
\item{data}{the dataframe holding your data}

\item{x}{X axis nutrient (will be labelled on an energy (kCal) basis)}

\item{y}{Y axis nutrient (will be labelled on an energy (kCal) basis)}

\item{z}{Z/implicit axis nutrient (will be labelled on an energy (kCal) basis)}

\item{id}{an optional argument. Grouping variable or other indicator of multiple groups within your dataset.}
}
\value{
dataframe of proportional nutrient data, along with right-angle mixture triangle for data
}
\description{
This function generates a right-angle mixture triangle demonstrating the nutrient balance of foods. 
All you need to do is input 3 nutrients and a grouping variable, and voila!
}
\examples{
RMTfunction(gorillafood,carb,fat,protein,age)
}
